Smooth the transition between different levels of granularity
This required rendering both map geometries at once (one on top of the other), because otherwise we have to wait for the new map to be drawn, which is slow.

There is a potential performance cost now because we are also updating both maps when we do "update map for year" so that is an area I would like to refactor soon.